refactor: create hexagonal hierarchy packages
