+++ src/main/scala/infrastructure/ui/PairPanel.scala
@@ -1,4 +1,4 @@
-package view
+package infrastructure.ui
 
 import java.awt.Color
 import javax.swing.BorderFactory
